all: Adjust windows perms in fs package
